+++ packages/dart/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `arrPick`: Add `arrPick` method
 - `getParsedInfoFromAddress`: Add `getParsedInfoFromAddress` method
+- `getSlug`: Add `getSlug` method
 
 ## 1.2.1 (2026--)